+++ CHANGELOG.md
@@ -1,3 +1,22 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Implement the initialized version of the widget
+
+## 0.0.2
+
+- Fix bug: Must highlight all same substrings instead of only highlighting the first occurence
+
+## 0.0.3
+
+- Add ignoreWordBoundaries option
+
+## 0.0.4
+
+- Fix bug: cannot identify word boundaries that relate to puntuations
+
+## 0.0.5
+
+- Add some reserved characters when ignoring diacritics
+## 0.0.6
+
+- Fix bug that the widget highlights a wrong range when having combining characters